Move object detection to folder
