+++ frigate/object_detection/base.py
