+++ frigate/types.py
@@ -3,7 +3,7 @@ from typing import TypedDict
 
 from frigate.camera import CameraMetrics
 from frigate.data_processing.types import DataProcessorMetrics
-from frigate.object_detection import ObjectDetectProcess
+from frigate.object_detection.base import ObjectDetectProcess
 
 
 class StatsTrackingTypes(TypedDict):
